Add missing `og:image:alt`
This seems to have existed and is in the published html, but wasn't in
the markdown used to (at some point) generate that html, nor was it in
the spec.

There are several posts around the web saying this isn't a useful
attribute, but it has existed so we should document it.