Update templates for consistent content generation
While we should eventually get rid of the php here, it was historically
being used to serve the site content based on the markdown.

This brings them back in sync, with many updates to prefer `https`. This
now makes it possible for any further updates to be made directly to the
markdown, and build the html from that with `php index.php >| index.html`