Update content sources for current state (#69)
* Add missing `og:image:alt`

This seems to have existed and is in the published html, but wasn't in
the markdown used to (at some point) generate that html, nor was it in
the spec.

There are several posts around the web saying this isn't a useful
attribute, but it has existed so we should document it.

* Update templates for consistent content generation

While we should eventually get rid of the php here, it was historically
being used to serve the site content based on the markdown.

This brings them back in sync, with many updates to prefer `https`. This
now makes it possible for any further updates to be made directly to the
markdown, and build the html from that with `php index.php >| index.html`