Use datatype string for profile:gender property
Change the datatype for the profile:gender property to string.
Previously, the datatype was enum(male, female), which made it effectively impossible to encode genders other than male or female.

In the EU, you are required by law (Article 16 GDPR, Right to rectification) to correct incorrect information.
Due to the standard, it was either not permitted to encode gender in some jurisdictions (e.g., Germany), where more genders than male or female are legally recognized, or you needed to ignore the standard.

The decision to use a string, instead of expanding the enum to include more gender descriptions, was based on the fact that different jurisdictions legally recognize different genders.
To keep the standard simple, this implementation was chosen.
Also, not explicitly mentioning non-binary genders will make this change less prone to attacks from TERFs, Trump supporters, or other assholes.

This is, first and foremost, a political commit. Any technical content you encounter is a byproduct of political acts and should be considered as such.